% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.transformNumeric}
\alias{.transformNumeric}
\title{Transformation of numeric}
\usage{
.transformNumeric(x, marker, tR = tR, U = U, L = L)
}
\arguments{
\item{x}{\code{numeric} migration time vector in seconds.}
}
\value{
\code{numeric} vector that represents effective mobility in mm^2 / (kV * min)
}
\description{
Transformation of numeric
}
\examples{
rtime <- c(10,20,30,40,50,60,70,80,90,100)
marker <- data.frame(markerID = c("marker1", "marker2"),
                     rtime = c(20,80),
                     mobility = c(0, 2000))
transformNumeric(x = rtime, marker = marker), tR = 3, U = 30, L = 90)

}
