% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.transformOnDiskMSnExp}
\alias{.transformOnDiskMSnExp}
\title{Transformation of OnDiskMSnExp}
\usage{
.transformOnDiskMSnExp(x, marker, tR = tR, U = U, L = L)
}
\arguments{
\item{x}{\code{OnDiskMSnExp}-object that stores the migration times in seconds.}
}
\value{
\code{OnDiskMSnExp}-Object that stores the effective mobility in mm^2 / (kV * min).
}
\description{
Transformation of OnDiskMSnExp
}
\examples{
fl <- system.file("extdata/CEMS_metabolites_10ppm_pos_centroidedData.mzML", 
package = "MobilityTransformR")
raw_data <- readMSData(files = fl,
                       mode = "onDisk")
marker <- data.frame(markerID = c("marker1", "marker2"),
                     rtime = c(20,80),
                     mobility = c(0, 2000),
                     fileIdx = c(1,1))
                     
.transformOnDiskMSnExp(x = raw_data, marker = marker)
}
