% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.transformSpectra}
\alias{.transformSpectra}
\title{Transformation of Spectra}
\usage{
.transformSpectra(x, marker, tR = tR, U = U, L = L)
}
\arguments{
\item{x}{\code{Spectra}-object that stores the migration times in seconds.}
}
\value{
\code{Spectra}-Object that stores the effective mobility in mm^2 / (kV * min).
}
\description{
Transformation of Spectra
}
\examples{
spectra_data <- Spectra(system.file("extdata/CEMS_10ppm.mzML", 
package = "MobilityTransformR"), backend = MsBackendMzR())
marker <- data.frame(markerID = c("marker1", "marker2"),
                     rtime = c(20,80),
                     mobility = c(0, 2000))
.transformSpectra(x = spectra_data, marker = marker, tR = 3, U = 30, L = 90)
}
