% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMt.R
\name{getMt}
\alias{getMt}
\title{get migration time of a known compound in CE-MS data}
\usage{
getMt(x, mz, mt, minInt, ...)
}
\arguments{
\item{x}{\code{OnDiskMSnExp} object, containing CE-MS(/MS) data from a single file (e.g.
filtered by \code{filterFile()}).
The migration time (mt) is provided in sec. The CE-MS data contain known
compounds \code{mz}, of which the exact migration time is going to be determined.}

\item{mz}{\code{numeric}, with upper and lower limit, (length(\code{mz}) = 2),
mz range of the compound of which the exact migration time
is going to be determined. The range should be as narrow as possible but
depends on the mass accuracy of the mass spectrometer that has been used
to acquire the data.}

\item{mt}{\code{numeric}, with upper and lower limit, (length(\code{mt}) = 2),
mt range of where the compound is expected to migrate.
Use a narrow \code{mt} range, to to avoid that other components with the same mz
and different mt are determined and to speed up the calculations.}

\item{minInt}{\code{numeric}, minimum intensity of the compound of which the exact migration
time is going to be determined. The \code{minInt} should be min at S/N = 3, to
ensure that no noise is beeing picked}
}
\value{
single \code{numeric} that represents the migration time in sec
}
\description{
The function \code{getMt} searches the migration time (mt) of a known compound
within a specified mz and mt range (\code{mz} and \code{mt}).
Also, a minimum intensity value \code{minInt} of the
compound must be provided to filter from noise.
}
\details{
\code{getMt} uses \code{OnDiskMSnExp} object of CE-MS data to search for the
migration time of selected compounds as for example EOF markers that where
included to the analysis to later perform effective mobility scale
transformation.
The \code{OnDiskMSnExp} object is beeing filtered using the defined mz-range \code{mz},
which depends on the mass accuracy of the mass spectrometer, and the
-range \code{mt}, where the compound is expected to migrate.
The migration time is determined by searching the mt of the maximum intense
data point in the produced extracted ion electropherogram. Therefore,
\code{minInt} is used to filter from noise.
}
\examples{
 ######  example to be added!

}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
