% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityTransform.R
\name{mobilityTransform}
\alias{mobilityTransform}
\title{Effective mobility scale transformation}
\usage{
mobilityTransform(
  tM,
  type = c("singleMarker", "twoMarker"),
  tEOF = NA,
  tA = NA,
  tR = 0,
  mobilityA = NA,
  U = NA,
  L = NA,
  l = L,
  ...
)
}
\arguments{
\item{tM}{\code{numeric} migration time \code{tM} in sec which will be transformed to an
effective mobility scale}

\item{type}{\code{character} defines if one or two markers are beeing used for effective
mobility scale transformation.
If \code{type} is \code{singleMarker} either \code{tEOF} or \code{tA} (in combination with
\code{mobilityA}) must be provided along with \code{U}, \code{l}, and \code{L}
If \code{type} is \code{twoMarker}, \code{tEOF}, \code{tA}, and \code{mobilityA} must be provided.

@param tEOF
\code{numeric} migration time of the (neutral charged) EOF marker in sec

@param tA
\code{numeric} migration time of the charged EOF marker in sec

@param tR
\code{numeric} field ramping delay in sec, default is 0

@param mobilityA
\code{numeric} effective mobility of the charged marker \code{tA} in mm^2 / (kV * min)

@param U
\code{numeric} applied electrical voltage U in kV. Note that in reversed CE
polarity, also a negative voltage is applied.

@param L
\code{numeric} total length of the capillary in mm

@param l
\code{numeric} effective length of the capillary in mm. In CE-MS l = L and in
CE-UV l < L. The default is l = L.}
}
\value{
\code{numeric} that represents effective mobility in mm^2 / (kV * min)
}
\description{
The function \code{mobilityTransform} performs effective mobility scale
transformation of CE data
}
\details{
\code{mobilityTransform} performs effective mobility scale transformation of
time-scaled CE data \code{tM}.
For the transformation, either one mobility marker (\code{type} = "singleMarker")
or both (\code{type} = "towMarker") were used.
If a single marker is used, either a neutral EOF marker \code{tEOF}, or charged
marker \code{tA} with its corresponding mobility \code{mobilityA} must be provided,
along with the applied voltage \code{U}, and the total and effective capillary
length \code{L} and \code{l}.
If two markers are used, both a neutral EOF marker \code{tEOF}, or charged
marker \code{tA} with its corresponding mobility \code{mobilityA} must be provided.
Additionally, field ramping delays can be included by \code{tR}, which will
result in more precise effective mobility values.
}
\examples{
 ######  example to be added!

}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
