% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformNumeric}
\alias{transformNumeric}
\title{Effective mobility scale transformation of a numeric migration time
vector}
\usage{
transformNumeric(x, marker, ...)
}
\arguments{
\item{x}{\code{numeric} migration time vector in seconds.}

\item{marker}{\code{data.frame} containing minimum two columns, where one holds the determined
migration time in minutes (here referred to as "rtime") of the EOF marker in
the same run in which the migration time is going to be transformed and the
other column the respective mobility ("mobility") of the EOF markers. Each
row hold the values for one EOF marker.
One or two entries are required for the transformation and depending on
the number of entries the transformation will be performed either on one or
both markers.

@param ...
Additional parameters passed to \code{convertMtime}, as for example \code{L}, \code{U},
and \code{tR}.}
}
\value{
\code{numeric} vector that represents effective mobility in mm^2 / (kV * min)
}
\description{
\code{transformNumeric} performs effective mobility scale transformation of a
\code{numeric} migration time vector as input. This can be used to transform a
row of migration times or a single value.
}
\details{
\code{transformNumeric} performs effective mobility scale transformation of
migration time vector.
For the transformation, either one mobility marker or both can be used.
If a single marker is used, either a neutral EOF marker, or charged
marker with its corresponding mobilities (0 for the neutral marker) must be
provided, along with the applied voltage \code{U}, and the total capillary length
\code{L}.
If two markers are used, both a neutral EOF marker and a charged marker
including their corresponding mobility must be provided.
Additionally, field ramping delays can be included by \code{tR}, which will
result in more precise effective mobility values.
}
\author{
Liesa Salzer
}
