% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformOnDiskMSnExp}
\alias{transformOnDiskMSnExp}
\title{Effective mobility scale transformation of a OnDiskMSnExp-object}
\usage{
transformOnDiskMSnExp(x, marker, ...)
}
\arguments{
\item{x}{\code{OnDiskMSnExp}-object that stores the migration times in seconds.}

\item{marker}{\code{data.frame} containing minimum two columns, where one holds the determined
migration time in minutes (here referred to as "rtime") of the EOF marker in
the same run in which the migration time is going to be transformed and the
other column the respective mobility ("mobility") of the EOF markers. Each
row hold the values for one EOF marker.
One or two entries are required for the transformation and depending on
the number of entries the transformation will be performed either on one or
both markers.

@param ...
Additional parameters passed to \code{convertMtime}, as for example \code{L}, \code{U},
and \code{tR}.}
}
\value{
\code{OnDiskMSnExp}-Object that stores the effective mobility in mm^2 / (kV * min).
Note that the effective mobility can be accessed by x$rtime
}
\description{
\code{transformOnDiskMSnExp} performs effective mobility scale transformation of the
migration time scale whithin an \code{OnDiskMSnExp} object.
}
\details{
\code{transformOnDiskMSnExp} performs effective mobility scale transformation of
migration times whithin an \code{OnDiskMSnExp}-object.
For the transformation, either one mobility marker or both can be used.
If a single marker is used, either a neutral EOF marker, or charged
marker with its corresponding mobilities (0 for the neutral marker) must be
provided, along with the applied voltage \code{U}, and the total capillary length
\code{L}.
If two markers are used, both a neutral EOF marker and a charged marker
including their corresponding mobility must be provided.
Additionally, field ramping delays can be included by \code{tR}, which will
result in more precise effective mobility values.
}
\author{
Liesa Salzer
}
