% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.transformNumeric}
\alias{.transformNumeric}
\title{Transformation of numeric}
\usage{
.transformNumeric(x, marker, tR = tR, U = U, L = L)
}
\arguments{
\item{x}{\code{numeric} migration time vector in seconds.}
}
\value{
\code{numeric} vector that represents effective mobility in mm^2 / (kV * min)
}
\description{
Transformation of numeric
}
