% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.transformOnDiskMSnExp}
\alias{.transformOnDiskMSnExp}
\title{Transformation of OnDiskMSnExp}
\usage{
.transformOnDiskMSnExp(x, marker, ...)
}
\arguments{
\item{x}{\code{OnDiskMSnExp}-object that stores the migration times in seconds.}
}
\value{
\code{OnDiskMSnExp}-Object that stores the effective mobility in mm^2 / (kV * min).
}
\description{
Transformation of OnDiskMSnExp
}
