% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.transformSpectra}
\alias{.transformSpectra}
\title{Transformation of Spectra}
\usage{
.transformSpectra(x, marker, ...)
}
\arguments{
\item{x}{\code{Spectra}-object that stores the migration times in seconds.}
}
\value{
\code{Spectra}-Object that stores the effective mobility in mm^2 / (kV * min).
}
\description{
Transformation of Spectra
}
