% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mobilityTransform}
\alias{mobilityTransform}
\title{Effective mobility scale transformation of CE-MS data}
\usage{
mobilityTransform(x, marker, ...)
}
\arguments{
\item{x}{\code{numeric} migration time vector, \code{Spectra} -object, or \code{MSnOnDiskExp}-object
that serves as input file to perform the effectibe mobility transformation.
The respective migration time scale should be in seconds(!).}

\item{marker}{\code{data.frame} containing minimum two columns, where one holds the determined
migration time in minutes (here referred to as "rtime") of the EOF marker in
the same run in which the migration time is going to be transformed and the
other column the respective mobility ("mobility") of the EOF markers. Each
row hold the values for one EOF marker.
One or two entries are required for the transformation and depending on
the number of entries the transformation will be performed either on one or
both markers.

@param ...
Additional parameters passed to \code{convertMtime}, as for example \code{L}, \code{U},
and \code{tR}.

@param FUN
\code{function} that is defined by the input-class and performs the actual
transformation}
}
\value{
The same class as the input class, i.e. if a \code{numeric} is used as input a
\code{numeric} that represents effective mobility will be the output.
If a \code{Spectra}-Object is the input, also a \code{Spectra}-Object with transformed
mobility scale will be returned. The same applies for \code{MSnOnDiskExp}-objects.
The respective unit for the effective mobility is mm^2 / (kV * min)
}
\description{
\code{mobilityTransform} performs effective mobility scale transformation of CE(-MS)
data, which is used to overcome variations of the migration times, caused by
differences in the Electroosmotic Flow (EOF) between different runs.
In order to monitor the EOF and perform the transformation, neutral or
charged EOF markers are spiked into the sample before analysis. The
information of the EOF markers (migration time and effective mobility) will
be then used to perform the  effective mobility transformation of the
migration time scale.
\code{mobilityTransform} is a method that used different functions to convert
CE-MS data, depending on the input class. Currently, \code{mobilityTransform}
supports a \code{numeric} vector of migration times as input, a \code{Spectra}-object
or \code{MSnOnDiskExp}-objects.
}
\details{
\code{mobilityTransform} performs effective mobility scale transformation of
migration time-scaled CE data.
For the transformation, either one mobility marker or both can be used.
If a single marker is used, either a neutral EOF marker, or charged
marker with its corresponding mobilities (0 for the neutral marker) must be
provided, along with the applied voltage \code{U}, and the total capillary length
\code{L}.
If two markers are used, both a neutral EOF marker and a charged marker
including their corresponding mobility must be provided.
Additionally, field ramping delays can be included by \code{tR}, which will
result in more precise effective mobility values.
}
\examples{
 rtime <- c(10,20,30,40,50,60,70,80,90,100)
 marker <- data.frame(markerID = c("marker1", "marker2"),
                      rtime = c(20,80),
                      mobility = c(0, 2000))
 mobilityTransform(x = rtime, marker = marker)

}
\author{
Liesa Salzer
}
